Renamed module folder from SigSurfaceModule to Map
