+++ frontend/src/modules/registerAllModules.ts
@@ -2,4 +2,4 @@ import "./MyModule2/registerModule";
 import "./MyModule/registerModule";
 import "./SimulationTimeSeries/registerModule";
 import "./InplaceVolumetrics/registerModule";
-import "./SigSurfaceModule/registerModule";
+import "./Map/registerModule";
